+++ panos/provider.go
@@ -65,6 +65,7 @@ func Provider() terraform.ResourceProvider {
 		ResourcesMap: map[string]*schema.Resource{
 			"panos_address_group":      resourceAddressGroup(),
 			"panos_address_object":     resourceAddressObject(),
+			"panos_administrative_tag": resourceAdministrativeTag(),
 			"panos_dag_tags":           resourceDagTags(),
 			"panos_ethernet_interface": resourceEthernetInterface(),
 			"panos_general_settings":   resourceGeneralSettings(),